Switched SCORE_TYPE to be double so that averaging doesn't suffer from integer truncation.

+++ src/common/depparser/implementations/arceager/depparser_macros.h
@@ -24,7 +24,7 @@ const unsigned DEP_LABEL_COUNT=CDependencyLabel::MAX_COUNT;
 const unsigned DEP_LABEL_COUNT = 1;
 #endif
 
-typedef int64_t SCORE_TYPE ;
+typedef double SCORE_TYPE ;
 #include "action.h"
 
 // supertag
